created post interface, restructurred the project folder
