+++ src/app/app.component.ts
@@ -5,7 +5,7 @@ import { Router } from '@angular/router';
 import { AngularFirestore } from '@angular/fire/compat/firestore';
 
 // Inject Auth service.
-import { AuthService } from './shared/services/auth.service';
+import { AuthService } from './shared/services/auth/auth.service';
 
 @Component({
 	selector: 'app-root',
