+++ src/app/app.module.ts
@@ -17,7 +17,7 @@ import { ForgotPasswordComponent } from './components/forgot-password/forgot-pas
 import { SignInComponent } from './components/sign-in/sign-in.component';
 import { SignUpComponent } from './components/sign-up/sign-up.component';
 import { VerifyEmailAddressComponent } from './components/verify-email-address/verify-email-address.component';
-import { AuthService } from './shared/services/auth.service';
+import { AuthService } from './shared/services/auth/auth.service';
 
 // Layout components.
 import { NavbarComponent } from './shared/components/navbar/navbar.component';
@@ -32,7 +32,7 @@ import { HomeComponent } from './components/home/home.component';
 		SignUpComponent,
 		VerifyEmailAddressComponent,
 		NavbarComponent,
-  HomeComponent,
+		HomeComponent,
 	],
 	imports: [
 		BrowserModule,
