+++ src/app/components/verify-email-address/verify-email-address.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 
 // Authentication services.
-import { AuthService } from 'src/app/shared/services/auth.service';
+import { AuthService } from 'src/app/shared/services/auth/auth.service';
 
 @Component({
 	selector: 'app-verify-email-address',
