+++ src/app/shared/guard/auth.guard.ts
@@ -5,7 +5,7 @@ import {
 	RouterStateSnapshot,
 	Router,
 } from '@angular/router';
-import { AuthService } from '../../shared/services/auth.service';
+import { AuthService } from '../services/auth/auth.service';
 import { Observable } from 'rxjs';
 @Injectable({
 	providedIn: 'root',
