+++ src/app/shared/services/auth/auth.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { AuthService } from './auth.service';
+
+describe('AuthService', () => {
+	let service: AuthService;
+
+	beforeEach(() => {
+		TestBed.configureTestingModule({});
+		service = TestBed.inject(AuthService);
+	});
+
+	it('should be created', () => {
+		expect(service).toBeTruthy();
+	});
+});
