+++ src/app/shared/services/models/post.ts
@@ -1,0 +1,9 @@
+export interface Post {
+	id: number;
+	slug: string;
+	backgroundImage: string;
+	title: string;
+	body: string;
+	displayName: string;
+	date: string;
+}
